+++ lib/kitchen/directions/bake_handbook/main.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Kitchen
+  module Directions
+    module BakeHandbook
+      def self.v1(book:, title_element: 'h1')
+        V1.new.bake(book: book, title_element: title_element)
+      end
+    end
+  end
+end